requirements/base.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
django-cors-headers==4.4.0
django-filter==24.2
psycopg2-binary==2.9.9
redis==5.0.8
celery==5.4.0
django-celery-beat==2.6.0
Pillow==10.4.0
gunicorn==22.0.0
drf-spectacular==0.27.2
